+++ packages/core/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact core module, newest chan
 ### Added
 
 - `core/handle.adaptEvent` to adapt event payload before passing to subsequent handler
+- `core/handle.call` to invoke a named method on a bound handler
 
 ## [2.0.0-alpha.7 - 2018-04-03]