+++ packages/core/CHANGELOG.md
@@ -2,14 +2,14 @@
 
 The following is a curated list of changes in the Enact core module, newest changes on the top.
 
-## [unreleased]
+## [2.0.0-alpha.8] - 2018-04-17
 
 ### Added
 
 - `core/handle.adaptEvent` to adapt event payload before passing to subsequent handler
 - `core/handle.call` to invoke a named method on a bound handler
 
-## [2.0.0-alpha.7 - 2018-04-03]
+## [2.0.0-alpha.7] - 2018-04-03
 
 No significant changes.